# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MH Interior Design API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface CustomerUser {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  role: string;
  isLocked: boolean;
  createdAt: string;
}

export interface RegisterInput {
  /** @minLength 2 */
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
  phone?: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface AuthResponse {
  user: CustomerUser;
  token: string;
}

export interface PasswordChangeInput {
  currentPassword: string;
  /** @minLength 6 */
  newPassword: string;
}

export interface ProfileInput {
  name?: string;
  phone?: string;
  email?: string;
}

export interface UserLockInput {
  isLocked: boolean;
}

export interface TimeSlot {
  time: string;
  label?: string;
  available: boolean;
}

export interface BookingInput {
  /** @minLength 2 */
  fullName: string;
  /** @minLength 10 */
  phone: string;
  email: string;
  service: string;
  preferredDate: string;
  timeSlot: string;
  budgetRange: string;
  projectLocation: string;
  /** @nullable */
  message?: string | null;
}

export interface Booking {
  id: number;
  consultationId: string;
  /** @nullable */
  userId?: number | null;
  fullName: string;
  phone: string;
  email: string;
  service: string;
  preferredDate: string;
  timeSlot: string;
  budgetRange: string;
  projectLocation: string;
  /** @nullable */
  message?: string | null;
  status: string;
  /** @nullable */
  qrCode?: string | null;
  createdAt: string;
}

export interface BookingVerification {
  booking: Booking;
  verified: boolean;
}

export interface BookingStatusInput {
  status: string;
}

export interface ReviewInput {
  /** @minLength 2 */
  name: string;
  /**
     * @minimum 1
     * @maximum 5
     */
  rating: number;
  /** @minLength 10 */
  message: string;
  /** @nullable */
  photoUrl?: string | null;
  /** @nullable */
  city?: string | null;
}

export interface Review {
  id: number;
  name: string;
  rating: number;
  message: string;
  /** @nullable */
  photoUrl?: string | null;
  /** @nullable */
  city?: string | null;
  status: string;
  isPinned: boolean;
  createdAt: string;
}

export interface ReviewStatusInput {
  status: string;
}

export interface ReviewPinInput {
  isPinned: boolean;
}

export interface ContactInput {
  /** @minLength 2 */
  name: string;
  phone: string;
  email: string;
  service: string;
  /** @minLength 5 */
  message: string;
}

export interface Contact {
  id: number;
  name: string;
  phone: string;
  email: string;
  service: string;
  message: string;
  isReplied: boolean;
  createdAt: string;
}

export interface AdminCredentials {
  password: string;
}

export interface AdminToken {
  token: string;
}

export interface AdminStatus {
  ok: boolean;
}

export interface AdminDashboard {
  totalBookings: number;
  pendingBookings: number;
  confirmedBookings: number;
  completedBookings: number;
  rejectedBookings: number;
  totalUsers: number;
  totalReviews: number;
  pendingReviews: number;
  unreadMessages: number;
  recentBookings: Booking[];
}

export interface ChatHistoryItem {
  role: string;
  content: string;
}

export interface ChatInput {
  /** @minLength 1 */
  message: string;
  history?: ChatHistoryItem[];
}

export interface ChatResponse {
  reply: string;
}

export type GetAvailableSlotsParams = {
date: string;
};

export type ListMyBookingsParams = {
status?: string;
search?: string;
};

export type ListBookingsParams = {
status?: string;
search?: string;
};

export type ListReviewsParams = {
status?: string;
};

export type ListContactsParams = {
search?: string;
replied?: boolean;
};