# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: MH Interior Design API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new customer account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Customer login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user profile
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerUser"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/change-password:
    post:
      operationId: changePassword
      tags: [auth]
      summary: Change password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordChangeInput"
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /auth/profile:
    patch:
      operationId: updateProfile
      tags: [auth]
      summary: Update profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileInput"
      responses:
        "200":
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerUser"

  # ── Bookings ────────────────────────────────────────────────────────────────
  /bookings/slots:
    get:
      operationId: getAvailableSlots
      tags: [bookings]
      summary: Get available time slots for a date
      parameters:
        - name: date
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of available time slots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TimeSlot"
  /bookings/my:
    get:
      operationId: listMyBookings
      tags: [bookings]
      summary: List current user's bookings
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: User's bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
  /bookings/verify/{consultationId}:
    get:
      operationId: verifyBooking
      tags: [bookings]
      summary: Verify booking by consultation ID (QR scan)
      parameters:
        - name: consultationId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Booking verification result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingVerification"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /bookings:
    get:
      operationId: listBookings
      tags: [bookings]
      summary: List all bookings (admin)
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
    post:
      operationId: createBooking
      tags: [bookings]
      summary: Create a new consultation booking (auth required)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingInput"
      responses:
        "201":
          description: Booking created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "409":
          description: Time slot already booked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /bookings/{id}/status:
    patch:
      operationId: updateBookingStatus
      tags: [bookings]
      summary: Update booking status (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingStatusInput"
      responses:
        "200":
          description: Status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"

  # ── Reviews ─────────────────────────────────────────────────────────────────
...
  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminCredentials"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminToken"
        "401":
          description: Invalid password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/users/{id}/lock:
    patch:
      operationId: lockUser
      tags: [admin]
      summary: Lock or unlock a user account (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserLockInput"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerUser"

  # ── Chat ─────────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]

    # ── Auth ──────────────────────────────────────────────────────────────────
    CustomerUser:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        role:
          type: string
        isLocked:
          type: boolean
        createdAt:
          type: string
      required: [id, name, email, role, isLocked, createdAt]
    RegisterInput:
      type: object
      properties:
        name:
          type: string
          minLength: 2
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        phone:
          type: string
      required: [name, email, password]
    LoginInput:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/CustomerUser"
        token:
          type: string
      required: [user, token]
    PasswordChangeInput:
      type: object
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
          minLength: 6
      required: [currentPassword, newPassword]
    ProfileInput:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: string
    UserLockInput:
      type: object
      properties:
        isLocked:
          type: boolean
      required: [isLocked]

    # ── Bookings ──────────────────────────────────────────────────────────────
    TimeSlot:
      type: object
      required: [time, available]
      properties:
        time:
          type: string
        label:
          type: string
        available:
          type: boolean
    BookingInput:
      type: object
      required: [fullName, phone, email, service, preferredDate, timeSlot, budgetRange, projectLocation]
      properties:
        fullName:
          type: string
          minLength: 2
        phone:
          type: string
          minLength: 10
        email:
          type: string
          format: email
        service:
          type: string
        preferredDate:
          type: string
        timeSlot:
          type: string
        budgetRange:
          type: string
        projectLocation:
          type: string
        message:
          type: ["string", "null"]
    Booking:
      type: object
      required: [id, consultationId, fullName, phone, email, service, preferredDate, timeSlot, budgetRange, projectLocation, status, createdAt]
      properties:
        id:
          type: integer
        consultationId:
          type: string
        userId:
          type: ["integer", "null"]
        fullName:
          type: string
        phone:
          type: string
        email:
          type: string
        service:
          type: string
        preferredDate:
          type: string
        timeSlot:
          type: string
        budgetRange:
          type: string
        projectLocation:
          type: string
        message:
          type: ["string", "null"]
        status:
          type: string
        qrCode:
          type: ["string", "null"]
        createdAt:
          type: string
    BookingVerification:
      type: object
      properties:
        booking:
          $ref: "#/components/schemas/Booking"
        verified:
          type: boolean
      required: [booking, verified]
    BookingStatusInput:
      type: object
      required: [status]
      properties:
        status:
          type: string

    # ── Reviews ───────────────────────────────────────────────────────────────
...
    AdminCredentials:
      type: object
      required: [password]
      properties:
        password:
          type: string
    AdminToken:
      type: object
      required: [token]
      properties:
        token:
          type: string
...