# lib/api-zod/src/generated/types/booking.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MH Interior Design API
 * OpenAPI spec version: 0.1.0
 */

export interface Booking {
  id: number;
  consultationId: string;
  /** @nullable */
  userId?: number | null;
  fullName: string;
  phone: string;
  email: string;
  service: string;
  preferredDate: string;
  timeSlot: string;
  budgetRange: string;
  projectLocation: string;
  /** @nullable */
  message?: string | null;
  status: string;
  /** @nullable */
  qrCode?: string | null;
  createdAt: string;
}