# lib/api-zod/src/generated/types/bookingInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MH Interior Design API
 * OpenAPI spec version: 0.1.0
 */

export interface BookingInput {
  /** @minLength 2 */
  fullName: string;
  /** @minLength 10 */
  phone: string;
  email: string;
  service: string;
  preferredDate: string;
  timeSlot: string;
  budgetRange: string;
  projectLocation: string;
  /** @nullable */
  message?: string | null;
}