# lib/api-zod/src/generated/types/customerUser.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MH Interior Design API
 * OpenAPI spec version: 0.1.0
 */

export interface CustomerUser {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  role: string;
  isLocked: boolean;
  createdAt: string;
}